requirements.txt
telethon
python-dotenv
urlextract
